Move contribution actions to feature folder
