+++ apps/app/src/features/contribution-graph/interfaces/supported-actions.ts
@@ -1,0 +1,15 @@
+import {
+  ACTION_COMMENT_CREATE,
+  ACTION_PAGE_CREATE,
+  ACTION_PAGE_DUPLICATE,
+  ACTION_PAGE_REVERT,
+  ACTION_PAGE_UPDATE,
+} from '~/interfaces/activity';
+
+export const ContributionGraphActions = {
+  ACTION_PAGE_CREATE,
+  ACTION_PAGE_UPDATE,
+  ACTION_PAGE_DUPLICATE,
+  ACTION_PAGE_REVERT,
+  ACTION_COMMENT_CREATE,
+} as const;
