+++ apps/app/src/features/contribution-graph/server/services/aggregation-service.ts
@@ -2,8 +2,9 @@ import type { Aggregate, PipelineStage } from 'mongoose';
 import mongoose from 'mongoose';
 
 import { getUTCMidnight } from '~/features/contribution-graph/utils/contribution-graph-utils';
-import { ContributionGraphActions } from '~/interfaces/activity';
 import Activity from '~/server/models/activity';
+
+import { ContributionGraphActions } from '../../interfaces/supported-actions';
 
 export interface PipelineParams {
   userId: string;
