+++ apps/app/src/interfaces/activity.ts
@@ -42,13 +42,13 @@ const ACTION_PAGE_LIKE = 'PAGE_LIKE';
 const ACTION_PAGE_UNLIKE = 'PAGE_UNLIKE';
 const ACTION_PAGE_BOOKMARK = 'PAGE_BOOKMARK';
 const ACTION_PAGE_UNBOOKMARK = 'PAGE_UNBOOKMARK';
-const ACTION_PAGE_CREATE = 'PAGE_CREATE';
-const ACTION_PAGE_UPDATE = 'PAGE_UPDATE';
+export const ACTION_PAGE_CREATE = 'PAGE_CREATE';
+export const ACTION_PAGE_UPDATE = 'PAGE_UPDATE';
 const ACTION_PAGE_RENAME = 'PAGE_RENAME';
-const ACTION_PAGE_DUPLICATE = 'PAGE_DUPLICATE';
+export const ACTION_PAGE_DUPLICATE = 'PAGE_DUPLICATE';
 const ACTION_PAGE_DELETE = 'PAGE_DELETE';
 const ACTION_PAGE_DELETE_COMPLETELY = 'PAGE_DELETE_COMPLETELY';
-const ACTION_PAGE_REVERT = 'PAGE_REVERT';
+export const ACTION_PAGE_REVERT = 'PAGE_REVERT';
 const ACTION_PAGE_EMPTY_TRASH = 'PAGE_EMPTY_TRASH';
 const ACTION_PAGE_RECURSIVELY_RENAME = 'PAGE_RECURSIVELY_RENAME';
 const ACTION_PAGE_RECURSIVELY_DELETE = 'PAGE_RECURSIVELY_DELETE';
@@ -64,7 +64,7 @@ const ACTION_PAGE_BULK_EXPORT_JOB_EXPIRED = 'PAGE_BULK_EXPORT_JOB_EXPIRED';
 const ACTION_TAG_UPDATE = 'TAG_UPDATE';
 const ACTION_IN_APP_NOTIFICATION_ALL_STATUSES_OPEN =
   'IN_APP_NOTIFICATION_ALL_STATUSES_OPEN';
-const ACTION_COMMENT_CREATE = 'COMMENT_CREATE';
+export const ACTION_COMMENT_CREATE = 'COMMENT_CREATE';
 const ACTION_COMMENT_UPDATE = 'COMMENT_UPDATE';
 const ACTION_COMMENT_REMOVE = 'COMMENT_REMOVE';
 const ACTION_SHARE_LINK_CREATE = 'SHARE_LINK_CREATE';
@@ -567,14 +567,6 @@ export const ActivityLogActions = {
   ACTION_PAGE_DELETE_COMPLETELY,
   ACTION_PAGE_REVERT,
   ACTION_PAGE_LIKE,
-  ACTION_COMMENT_CREATE,
-} as const;
-
-export const ContributionGraphActions = {
-  ACTION_PAGE_CREATE,
-  ACTION_PAGE_UPDATE,
-  ACTION_PAGE_DUPLICATE,
-  ACTION_PAGE_REVERT,
   ACTION_COMMENT_CREATE,
 } as const;
